Quote Database + SyncAdapter + Quote displayed in Main Activity view
Created a database for the storage of quotes. A SyncAdapter is used to
obtain quotes from the theysaidso.com API. The quotes should be
retrieved on a daily basis. The MainActivity queries the database
through the use of Loaders and displays the most recent quote in its
view.  Currently, the quote view appears to be consistently showing up
on my tablet, but I have trouble viewing it on the emulator. This code
has a lot of unnecessary logging statements, and it looks rather rough
due to my attempts at debugging (I apologize for that!). There are
certain necessary sections of code that are currently commented out
(such as the deletion of older quotes from the database), since I am
unsure of whether they are interfering with the proper insertion of
quotes into the database. Also, the refresh button (which tells the sync
adapter to sync immediately) should be removed,  but I haven't done that
yet since I think it'll be useful to future debugging (we can probably
remove it later when we're sure everything works properly!).